+++ playground/compiler-react-18/vite.config.ts
@@ -1,8 +1,6 @@
 import { defineConfig } from 'vite'
 import react from '@vitejs/plugin-react'
 
-// https://gist.github.com/poteto/37c076bf112a07ba39d0e5f0645fec43
-
 // https://vitejs.dev/config/
 export default defineConfig(({ command }) => {
   return {
@@ -10,12 +8,7 @@ export default defineConfig(({ command }) => {
     plugins: [
       react({
         babel: {
-          plugins: [
-            [
-              'babel-plugin-react-compiler',
-              { runtimeModule: 'react-compiler-runtime' },
-            ],
-          ],
+          plugins: [['babel-plugin-react-compiler', { target: '18' }]],
         },
       }),
     ],